testing: add test for top gainers list using mock data and mock Auth0 authenticated user
