+++ app/package.json
@@ -46,6 +46,7 @@
     "@testing-library/user-event": "^13.0.2",
     "eslint-config-prettier": "^8.2.0",
     "eslint-plugin-prettier": "^4.0.0",
+    "jest": "^26.6.0",
     "node-sass": "^6.0.1",
     "prettier": "^2.2.1"
   }
